requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
cloudinary==2.22.0
werkzeug==2.3.7
gunicorn==20.1.0
